set up prepend for redis
