+++ lib/new_relic/agent/configuration/default_source.rb
@@ -16,13 +16,13 @@ module NewRelic
       end
 
       # Combines potentially two properties into one.
-      # Given the example: 
+      # Given the example:
       #    :disable_net_http and :prepend_net_instrumentation
       #    if :disable_net_http is true, then returned value is "disabled"
       #    if :prepend_net_instrumentation is false, then returned value is "chain"
       #    otherwise, "auto" is returned.
       #
-      # Intent is: 
+      # Intent is:
       #     - if user sets disable_xxx property, then don't instrument
       #     - if user set prepend to `false` then we use method_alias chaining
       #     - auto, when returned means, try to use prepend unless conflicting gems discovered
@@ -906,6 +906,14 @@ module NewRelic
           :allowed_from_server => false,
           :description => "Controls auto-instrumentation of HTTPClient at start up.  May be one of [auto|prepend|chain|disabled]."
         },
+        :'instrumentation.redis_instrumentation' => {
+          :default => instrumentation_value_of(:disable_redis),
+          :public => :true,
+          :type => String,
+          :dynamic_name => true,
+          :allowed_from_server => false,
+          :description => "Controls auto-instrumentation of redis at start up.  May be one of [auto|prepend|chain|disabled]."
+        },
         :disable_data_mapper => {
           :default => false,
           :public => true,
@@ -1112,14 +1120,14 @@ module NewRelic
           :public       => true,
           :type         => Boolean,
           :allowed_from_server => false,
-          :description  => 'If <code>true</code>, the agent won\'t install <a href="https://docs.newrelic.com/docs/agents/ruby-agent/frameworks/redis-instrumentation">instrumentation for Redis</a>.'
+          :description  => deprecated_description(:'instrumentation.redis', 'If <code>true</code>, the agent won\'t install <a href="https://docs.newrelic.com/docs/agents/ruby-agent/frameworks/redis-instrumentation">instrumentation for Redis</a>.')
         },
         :disable_redis_instrumentation => {
           :default      => false,
           :public       => false,
           :type         => Boolean,
           :allowed_from_server => false,
-          :description  => 'Disables installation of Redis instrumentation. Standard key to use is disable_redis.'
+          :description  => deprecated_description(:'instrumentation.redis', 'Disables installation of Redis instrumentation. Standard key to use is disable_redis.')
         },
         :'message_tracer.segment_parameters.enabled' => {
           :default      => true,
@@ -1528,7 +1536,7 @@ module NewRelic
           :dynamic_name => true,
           :allowed_from_server => false,
           :deprecated   => true,
-          :description  => deprecated_description(:'instrumentation.net_http', 
+          :description  => deprecated_description(:'instrumentation.net_http',
             'If <code>true</code>, disables instrumentation for Net::HTTP.'
           )
         },
