+++ test/multiverse/suites/redis/config/newrelic.yml
@@ -7,6 +7,8 @@ development:
   license_key: bootstrap_newrelic_admin_license_key_000
   app_name: test
   ca_bundle_path: ../../../config/test.cert.crt
+  instrumentation:
+    redis_instrumentation: <%= $instrumentation_method %>
   app_name: test
   host: localhost
   api_host: localhost
